@@ -4,26 +4,28 @@
   "license": "MIT",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve",
+    "start": "ng serve --host 0.0.0.0",
     "build": "ng build --configuration production",
     "build-travis": "ng build --configuration production --output-path=$BUILD_UI_TRAVIS_DIR",
     "test": "ng test --watch=false",
     "lint": "ng lint",
-    "e2e": "ng e2e desktopwallet-e2e"
+    "e2e": "wdio run ./wdio.conf.ts --autoCompileOpts.tsNodeOpts.project=./e2e/tsconfig.json",
+    "e2e-with-server": "concurrently \"wait-on http-get://localhost:4200 && npm run e2e\" \"npm start\" -s first --kill-others"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^12.0.2",
-    "@angular/cdk": "^12.0.2",
-    "@angular/common": "^12.0.2",
-    "@angular/compiler": "^12.0.2",
-    "@angular/core": "^12.0.2",
-    "@angular/forms": "^12.0.2",
-    "@angular/material": "^12.0.2",
-    "@angular/platform-browser": "^12.0.2",
-    "@angular/platform-browser-dynamic": "^12.0.2",
-    "@angular/router": "^12.0.2",
-    "@ngx-translate/core": "^10.0.2",
+    "@angular/animations": "^16.0.2",
+    "@angular/cdk": "^16.0.1",
+    "@angular/common": "^16.0.2",
+    "@angular/compiler": "^16.0.2",
+    "@angular/core": "^16.0.2",
+    "@angular/forms": "^16.0.2",
+    "@angular/material": "^16.0.1",
+    "@angular/platform-browser": "^16.0.2",
+    "@angular/platform-browser-dynamic": "^16.0.2",
+    "@angular/router": "^16.0.2",
+    "@ngx-translate/core": "^15.0.0",
+    "@ngx-translate/http-loader": "^8.0.0",
     "base-x": "3.0.5",
     "bignumber.js": "^7.2.1",
     "bootstrap": "^4.0.0-beta.2",
@@ -32,29 +34,46 @@
     "moment": "^2.21.0",
     "rxjs": "^6.5.3",
     "tslib": "^2.0.0",
-    "zone.js": "~0.11.4"
+    "zone.js": "~0.13.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.0.2",
-    "@angular/cli": "^12.0.2",
-    "@angular/compiler-cli": "^12.0.2",
-    "@angular/language-service": "^12.0.2",
+    "@angular-devkit/build-angular": "^16.0.2",
+    "@angular-eslint/builder": "16.0.2",
+    "@angular-eslint/eslint-plugin": "16.0.2",
+    "@angular-eslint/eslint-plugin-template": "16.0.2",
+    "@angular-eslint/schematics": "16.0.2",
+    "@angular-eslint/template-parser": "16.0.2",
+    "@angular/cli": "^16.0.2",
+    "@angular/compiler-cli": "^16.0.2",
+    "@angular/language-service": "^16.0.2",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.1",
-    "codelyzer": "^6.0.0",
-    "jasmine-core": "~3.6.0",
+    "@typescript-eslint/eslint-plugin": "^5.59.6",
+    "@typescript-eslint/parser": "^5.59.6",
+    "@wdio/cli": "^8.3.5",
+    "@wdio/jasmine-framework": "^8.3.5",
+    "@wdio/local-runner": "^8.3.5",
+    "@wdio/spec-reporter": "^8.3.0",
+    "chromedriver": "^114.0.0",
+    "concurrently": "8.0.1",
+    "eslint": "^8.39.0",
+    "eslint-plugin-import": "2.25.4",
+    "eslint-plugin-jsdoc": "37.9.4",
+    "eslint-plugin-prefer-arrow": "1.2.3",
+    "jasmine-core": "~3.8.0",
     "jasmine-spec-reporter": "~5.0.0",
     "karma": "~6.3.2",
     "karma-chrome-launcher": "~3.1.0",
-    "karma-cli": "~2.0.0",
-    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-cli": "~1.0.1",
+    "karma-coverage": "~2.2.0",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
-    "karma-read-json": "^1.1.0",
     "protractor": "~7.0.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~6.1.0",
-    "typescript": "~4.2.4"
+    "ts-node": "^10.9.1",
+    "typescript": "~5.0.4",
+    "wait-on": "6.0.1",
+    "wdio-chromedriver-service": "^8.0.1",
+    "wdio-wait-for": "^3.0.2"
   }
 }
