@@ -1,7 +1,8 @@
-import { Component, OnDestroy, ViewChild, ChangeDetectorRef } from '@angular/core';
+import { Component, OnDestroy, ViewChild, ChangeDetectorRef, OnInit } from '@angular/core';
 import { SubscriptionLike, combineLatest } from 'rxjs';
 import { ActivatedRoute, Router } from '@angular/router';
-import { FormGroup, FormBuilder, FormControl } from '@angular/forms';
+import { UntypedFormGroup, UntypedFormBuilder, UntypedFormControl } from '@angular/forms';
+import { MatDialog } from '@angular/material/dialog';
 import { map } from 'rxjs/operators';
 
 import { ButtonComponent } from '../../layout/button/button.component';
@@ -9,6 +10,9 @@ import { MsgBarService } from '../../../services/msg-bar.service';
 import { SoftwareWalletService } from '../../../services/wallet-operations/software-wallet.service';
 import { WalletsAndAddressesService } from '../../../services/wallet-operations/wallets-and-addresses.service';
 import { WalletBase } from '../../../services/wallet-operations/wallet-objects';
+import { AssistedSeedFieldComponent } from '../wallets/create-wallet/create-wallet-form/assisted-seed-field/assisted-seed-field.component';
+import { WordAskedReasons } from '../../layout/seed-word-dialog/seed-word-dialog.component';
+import { ConfirmationComponent, ConfirmationParams, DefaultConfirmationButtons } from '../../layout/confirmation/confirmation.component';
 
 /**
  * Allows to use the seed to remove or change the password of an encrypted software wallet.
@@ -20,31 +24,43 @@ import { WalletBase } from '../../../services/wallet-operations/wallet-objects';
   templateUrl: './reset-password.component.html',
   styleUrls: ['./reset-password.component.scss'],
 })
-export class ResetPasswordComponent implements OnDestroy {
+export class ResetPasswordComponent implements OnInit, OnDestroy {
   @ViewChild('resetButton') resetButton: ButtonComponent;
+  // Component for entering the seed using the assisted mode.
+  @ViewChild('assistedSeed') assistedSeed: AssistedSeedFieldComponent;
 
-  form: FormGroup;
+  form: UntypedFormGroup;
   // Allows to deactivate the form while the component is busy.
   busy = true;
+  // If the id on the URL does not correspond to a valid wallet.
+  invalidWallet = false;
+  // If true, the user must enter the seed using the asisted mode.
+  enterSeedWithAssistance = true;
+
+  wordAskedReasons = WordAskedReasons;
 
   // Vars with the validation error messages.
   seedErrorMsg = '';
   passwordErrorMsg = '';
 
+  wallet: WalletBase;
+
   private subscription: SubscriptionLike;
-  private wallet: WalletBase;
   private done = false;
   private hideBarWhenClosing = true;
 
   constructor(
-    public formBuilder: FormBuilder,
+    public formBuilder: UntypedFormBuilder,
     private route: ActivatedRoute,
     private router: Router,
     private msgBarService: MsgBarService,
     private softwareWalletService: SoftwareWalletService,
     private walletsAndAddressesService: WalletsAndAddressesService,
     private changeDetector: ChangeDetectorRef,
-  ) {
+    private dialog: MatDialog,
+  ) { }
+
+    ngOnInit() {
     this.initForm();
     // Get the wallets and route params.
     this.subscription = combineLatest([this.route.params, this.walletsAndAddressesService.allWallets]).pipe(map(result => {
@@ -52,9 +68,11 @@ export class ResetPasswordComponent implements OnDestroy {
       const wallets = result[1];
 
       const wallet = wallets.find(w => w.id === params['id']);
-      // Abort if the requested wallet does not exists or is temporal.
-      if (!wallet || wallet.temporal) {
-        setTimeout(() => this.router.navigate([''], {skipLocationChange: true}));
+      this.invalidWallet = false;
+
+      // Abort if the requested wallet does not exists.
+      if (!wallet) {
+        this.invalidWallet = true;
 
         return;
       }
@@ -73,14 +91,56 @@ export class ResetPasswordComponent implements OnDestroy {
     }
   }
 
+  // Returns the value of the number_of_words form field.
+  get selectedNumberOfWords(): number {
+    return this.form ? this.form.value.number_of_words : 0;
+  }
+
   initForm() {
-    this.form = new FormGroup({});
-    this.form.addControl('wallet', new FormControl());
-    this.form.addControl('seed', new FormControl(''));
-    this.form.addControl('password', new FormControl(''));
-    this.form.addControl('confirm', new FormControl(''));
+    this.form = new UntypedFormGroup({});
+    this.form.addControl('wallet', new UntypedFormControl());
+    this.form.addControl('number_of_words', new UntypedFormControl(12));
+    this.form.addControl('seed', new UntypedFormControl(''));
+    this.form.addControl('password', new UntypedFormControl(''));
+    this.form.addControl('confirm', new UntypedFormControl(''));
+
+    // If assistedSeed already exists, reset it.
+    if (this.assistedSeed) {
+      this.assistedSeed.lastAssistedSeed = '';
+    }
 
     this.form.setValidators(this.validateForm.bind(this));
+  }
+
+  // Switches between the assisted mode and the manual mode for entering the seed.
+  changeSeedType() {
+    this.enterSeedWithAssistance = !this.enterSeedWithAssistance;
+    this.form.updateValueAndValidity();
+  }
+
+  assistedSeedChanged() {
+    this.form.updateValueAndValidity();
+  }
+
+  // Checks if the password will be removed and asks for confirmation before continuing to reset().
+  checkBeforeReset() {
+    if (this.form.value.password) {
+      this.reset();
+    } else {
+      const confirmationParams: ConfirmationParams = {
+        headerText: 'reset.remove-password-confirmation-title',
+        redTitle: true,
+        text: 'reset.remove-password-confirmation',
+        defaultButtons: DefaultConfirmationButtons.YesNo,
+      };
+
+      // Ask for confirmation.
+      ConfirmationComponent.openDialog(this.dialog, confirmationParams).afterClosed().subscribe(confirmationResult => {
+        if (confirmationResult) {
+          this.reset();
+        }
+      });
+    }
   }
 
   // Resets the wallet password.
@@ -93,7 +153,9 @@ export class ResetPasswordComponent implements OnDestroy {
     this.msgBarService.hide();
     this.resetButton.setLoading();
 
-    this.softwareWalletService.resetPassword(this.wallet, this.form.value.seed, this.form.value.password !== '' ? this.form.value.password : null)
+    const seed = this.enterSeedWithAssistance ? this.assistedSeed.lastAssistedSeed : this.form.value.seed;
+
+    this.softwareWalletService.resetPassword(this.wallet, seed, this.form.value.password !== '' ? this.form.value.password : null)
       .subscribe(() => {
         this.resetButton.setSuccess();
         this.resetButton.setDisabled();
@@ -127,10 +189,16 @@ export class ResetPasswordComponent implements OnDestroy {
 
     let valid = true;
 
-    if (!this.form.get('seed').value) {
-      valid = false;
-      if (this.form.get('seed').touched) {
-        this.seedErrorMsg = 'reset.seed-error-info';
+    if (this.enterSeedWithAssistance) {
+      if (!this.assistedSeed || !this.assistedSeed.lastAssistedSeed) {
+        valid = false;
+      }
+    } else {
+      if (!this.form.get('seed').value) {
+        valid = false;
+        if (this.form.get('seed').touched) {
+          this.seedErrorMsg = 'reset.seed-error-info';
+        }
       }
     }
 
