@@ -1,5 +1,5 @@
 import { Component, OnInit, Inject, ViewChild, OnDestroy, ChangeDetectorRef } from '@angular/core';
-import { FormBuilder, FormGroup } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormGroup } from '@angular/forms';
 import { MAT_DIALOG_DATA, MatDialogRef, MatDialog, MatDialogConfig } from '@angular/material/dialog';
 import { SubscriptionLike } from 'rxjs';
 
@@ -66,7 +66,7 @@ export class ChangeNameErrorResponse {
 export class ChangeNameComponent implements OnInit, OnDestroy {
   // Confirmation button.
   @ViewChild('button') button: ButtonComponent;
-  form: FormGroup;
+  form: UntypedFormGroup;
   currentState: States = States.Initial;
   states = States;
   msgIcons = MessageIcons;
@@ -95,7 +95,7 @@ export class ChangeNameComponent implements OnInit, OnDestroy {
   constructor(
     public dialogRef: MatDialogRef<ChangeNameComponent>,
     @Inject(MAT_DIALOG_DATA) public data: ChangeNameData,
-    private formBuilder: FormBuilder,
+    private formBuilder: UntypedFormBuilder,
     private hwWalletService: HwWalletService,
     private msgBarService: MsgBarService,
     private softwareWalletService: SoftwareWalletService,
@@ -215,7 +215,7 @@ export class ChangeNameComponent implements OnInit, OnDestroy {
 
     let valid = true;
 
-    if (!this.form.get('label').value || (this.form.get('label').value as String).trim().length === 0) {
+    if (!this.form.get('label').value || (this.form.get('label').value as string).trim().length === 0) {
       valid = false;
       if (this.form.get('label').touched) {
         this.inputErrorMsg = 'wallet.rename.label-error-info';
