@@ -1,16 +1,17 @@
 import { switchMap, delay, flatMap } from 'rxjs/operators';
 import { Component, OnInit, OnDestroy, Input, Output, EventEmitter, ViewChild } from '@angular/core';
-import { FormControl, FormGroup } from '@angular/forms';
+import { UntypedFormControl, UntypedFormGroup } from '@angular/forms';
 import { SubscriptionLike, Subject, of } from 'rxjs';
 import { MatDialog } from '@angular/material/dialog';
+import { MatCheckbox } from '@angular/material/checkbox';
 
 import { ApiService } from '../../../../../services/api.service';
-import { SeedWordDialogComponent, WordAskedReasons } from '../../../../layout/seed-word-dialog/seed-word-dialog.component';
+import { WordAskedReasons } from '../../../../layout/seed-word-dialog/seed-word-dialog.component';
 import { MsgBarService } from '../../../../../services/msg-bar.service';
 import { ConfirmationParams, ConfirmationComponent, DefaultConfirmationButtons } from '../../../../layout/confirmation/confirmation.component';
 import { OperationError } from '../../../../../utils/operation-error';
 import { processServiceError } from '../../../../../utils/errors';
-import { MatCheckbox } from '@angular/material/checkbox';
+import { AssistedSeedFieldComponent } from './assisted-seed-field/assisted-seed-field.component';
 
 /**
  * Data entered in an instance of CreateWalletFormComponent.
@@ -37,8 +38,7 @@ export class WalletFormData {
    */
   enterSeedWithAssistance: boolean;
   /**
-   * If creating a new wallet, the last automatically generated seed for the assisted mode. If
-   * loading a wallet, the last valid seed the user entered using the assisted procedure.
+   * If creating a new wallet, the last automatically generated seed for the assisted mode.
    */
   lastAssistedSeed: string;
   /**
@@ -75,6 +75,8 @@ export class WalletFormData {
 })
 export class CreateWalletFormComponent implements OnInit, OnDestroy {
   @ViewChild('temporalWalletCheck') temporalWalletCheck: MatCheckbox;
+  // Component for entering the seed using the assisted mode.
+  @ViewChild('assistedSeed') assistedSeed: AssistedSeedFieldComponent;
 
   // If the form is for creating a new wallet (true) or loading a walled using a seed (false).
   @Input() create: boolean;
@@ -85,11 +87,9 @@ export class CreateWalletFormComponent implements OnInit, OnDestroy {
   // Emits when the user asks for the wallet ot be created.
   @Output() createRequested = new EventEmitter<void>();
 
-  form: FormGroup;
+  form: UntypedFormGroup;
   // If true, the user must enter the ssed using the asisted mode.
   enterSeedWithAssistance = true;
-  // If the user confirmed the seed using the asisted mode, while creating a new wallet.
-  assistedSeedConfirmed = false;
   // If the user entered a standard seed using the manual mode.
   customSeedIsNormal = true;
   // If the user entered a non-standard seed using the manual mode and confirmed to use it.
@@ -100,15 +100,13 @@ export class CreateWalletFormComponent implements OnInit, OnDestroy {
   showAdvancedOptions = false;
   // If the user selected that the wallet must be loaded temporarily.
   loadTemporarily = false;
-  // If creating a new wallet, the last automatically generated seed for the assisted mode. If
-  // loading a wallet, the last valid seed the user entered using the assisted procedure.
+  // If creating a new wallet, the last automatically generated seed for the assisted mode.
   lastAssistedSeed = '';
   // How many words the last autogenerated seed for the assisted mode has, when creating
   // a new wallet.
   numberOfAutogeneratedWords = 0;
   // If the system is currently checking the custom seed entered by the user.
   checkingCustomSeed = false;
-  checkingAssistedSeed = false;
 
   // Vars with the validation error messages.
   labelErrorMsg = '';
@@ -117,14 +115,13 @@ export class CreateWalletFormComponent implements OnInit, OnDestroy {
   password1ErrorMsg = '';
   password2ErrorMsg = '';
 
+  wordAskedReasons = WordAskedReasons;
+
   // Emits every time the seed should be checked again, to know if it is a standard seed.
   private seed: Subject<string> = new Subject<string>();
 
   private statusSubscription: SubscriptionLike;
   private seedValiditySubscription: SubscriptionLike;
-
-  // Saves the words the user enters while using the assisted mode.
-  private partialSeed: string[];
 
   constructor(
     private apiService: ApiService,
@@ -155,8 +152,7 @@ export class CreateWalletFormComponent implements OnInit, OnDestroy {
     return this.form.valid && !this.checkingCustomSeed &&
       (
         (!this.enterSeedWithAssistance && (this.customSeedIsNormal || this.customSeedAccepted)) ||
-        (this.create && this.enterSeedWithAssistance && this.assistedSeedConfirmed) ||
-        (!this.create && this.enterSeedWithAssistance && this.lastAssistedSeed.length > 2)
+        (this.enterSeedWithAssistance && this.assistedSeed.lastAssistedSeed)
       );
   }
 
@@ -213,7 +209,7 @@ export class CreateWalletFormComponent implements OnInit, OnDestroy {
     return {
       creatingNewWallet: this.create,
       label: this.form.value.label,
-      seed: this.enterSeedWithAssistance ? this.lastAssistedSeed : this.form.value.seed,
+      seed: this.enterSeedWithAssistance ? this.assistedSeed.lastAssistedSeed : this.form.value.seed,
       password: !this.onboarding && this.encrypt ? this.form.value.password : null,
       enterSeedWithAssistance: this.enterSeedWithAssistance,
       lastAssistedSeed: this.lastAssistedSeed,
@@ -251,82 +247,6 @@ export class CreateWalletFormComponent implements OnInit, OnDestroy {
     }
   }
 
-  // Starts the assisted procedure for entering the seed, if the user is trying to load
-  // an existing wallet.
-  enterSeed() {
-    if (!this.create) {
-      this.partialSeed = [];
-      this.askForWord(0);
-      this.msgBarService.hide();
-    }
-  }
-
-  // Starts the assisted procedure for confirming the automatically generated seed, if the
-  // user is trying to create a new wallet.
-  confirmSeed() {
-    if (!this.assistedSeedConfirmed) {
-      this.partialSeed = [];
-      this.askForWord(0);
-      this.msgBarService.hide();
-    }
-  }
-
-  /**
-   * Recursively asks the user to enter the words of the seed.
-   * @param wordIndex Index of the word which is going to be requested on this step. Must be
-   * 0 when starting to ask for the words.
-   */
-  private askForWord(wordIndex: number) {
-    // Open the modal window for entering the seed word.
-    return SeedWordDialogComponent.openDialog(this.dialog, {
-      reason: this.create ? WordAskedReasons.CreatingSoftwareWallet : WordAskedReasons.RecoveringSoftwareWallet,
-      wordNumber: wordIndex + 1,
-    }).afterClosed().subscribe(word => {
-      if (word) {
-        // If creating a new wallet, check if the user entered the requested word.
-        if (this.create) {
-          const lastSeedWords = this.lastAssistedSeed.split(' ');
-          if (word !== lastSeedWords[wordIndex]) {
-            this.msgBarService.showError('wallet.new.seed.incorrect-word-error');
-
-            return;
-          }
-        }
-
-        // Add the entered word to the list of words the user already entered.
-        this.partialSeed[wordIndex] = word;
-        wordIndex += 1;
-
-        if ((this.create && wordIndex < this.numberOfAutogeneratedWords) || (!this.create && wordIndex < this.form.controls['number_of_words'].value)) {
-          this.askForWord(wordIndex);
-        } else {
-          if (this.create) {
-            // Set the seed as confirmed.
-            this.assistedSeedConfirmed = true;
-          } else {
-            // Build the seed.
-            let enteredSeed = '';
-            this.partialSeed.forEach(currentWord => enteredSeed += currentWord + ' ');
-            enteredSeed = enteredSeed.substr(0, enteredSeed.length - 1);
-
-            // Check the seed and use it only if it is valid.
-            this.checkingAssistedSeed = true;
-            this.apiService.post('wallet/seed/verify', {seed: enteredSeed}, {useV2: true}).subscribe(
-              () => {
-                this.checkingAssistedSeed = false;
-                this.lastAssistedSeed = enteredSeed;
-              },
-              () => {
-                this.checkingAssistedSeed = false;
-                this.msgBarService.showError('wallet.new.seed.invalid-seed-error');
-              },
-            );
-          }
-        }
-      }
-    });
-  }
-
   /**
    * Inits or resets the form.
    * @param create If the form is for creating a new wallet (true) or loading a walled using
@@ -341,13 +261,13 @@ export class CreateWalletFormComponent implements OnInit, OnDestroy {
     this.lastAssistedSeed = '';
     this.enterSeedWithAssistance = true;
 
-    this.form = new FormGroup({});
-    this.form.addControl('label', new FormControl(data ? data.label : ''));
-    this.form.addControl('seed', new FormControl(data ? data.lastCustomSeed : ''));
-    this.form.addControl('confirm_seed', new FormControl(data ? data.lastCustomSeed : ''));
-    this.form.addControl('password', new FormControl());
-    this.form.addControl('confirm_password', new FormControl());
-    this.form.addControl('number_of_words', new FormControl(!this.create && data && data.numberOfWords ? data.numberOfWords : 12));
+    this.form = new UntypedFormGroup({});
+    this.form.addControl('label', new UntypedFormControl(data ? data.label : ''));
+    this.form.addControl('seed', new UntypedFormControl(data ? data.lastCustomSeed : ''));
+    this.form.addControl('confirm_seed', new UntypedFormControl(data ? data.lastCustomSeed : ''));
+    this.form.addControl('password', new UntypedFormControl());
+    this.form.addControl('confirm_password', new UntypedFormControl());
+    this.form.addControl('number_of_words', new UntypedFormControl(!this.create && data && data.numberOfWords ? data.numberOfWords : 12));
 
     this.form.setValidators(this.validateForm.bind(this));
 
@@ -362,7 +282,7 @@ export class CreateWalletFormComponent implements OnInit, OnDestroy {
     if (data) {
       this.enterSeedWithAssistance = data.enterSeedWithAssistance;
       this.lastAssistedSeed = data.lastAssistedSeed;
-      this.assistedSeedConfirmed = true;
+      this.assistedSeed.lastAssistedSeed = create ? data.lastAssistedSeed : data.seed;
       this.customSeedAccepted = true;
       this.customSeedIsNormal = data.customSeedIsNormal;
 
@@ -394,7 +314,7 @@ export class CreateWalletFormComponent implements OnInit, OnDestroy {
       this.numberOfAutogeneratedWords = 24;
     }
 
-    this.apiService.get('wallet/newSeed', { entropy }).subscribe(response => {
+    this.apiService.get('wallet/newSeed', { entropy: entropy }).subscribe(response => {
       this.lastAssistedSeed = response.seed;
       this.form.get('seed').setValue(response.seed);
       this.form.get('seed').markAsTouched();
@@ -414,7 +334,9 @@ export class CreateWalletFormComponent implements OnInit, OnDestroy {
    */
   private removeConfirmations(cleanSecondSeedField = true) {
     this.customSeedAccepted = false;
-    this.assistedSeedConfirmed = false;
+    if (this.assistedSeed) {
+      this.assistedSeed.lastAssistedSeed = null;
+    }
     if (cleanSecondSeedField) {
       this.form.get('confirm_seed').setValue('');
     }
@@ -432,7 +354,7 @@ export class CreateWalletFormComponent implements OnInit, OnDestroy {
       if (!this.enterSeedWithAssistance && seed.trim().length > 0 && (!this.create || this.seedsAreEqual())) {
         this.checkingCustomSeed = true;
 
-        return of(0).pipe(delay(500), flatMap(() => this.apiService.post('wallet/seed/verify', {seed}, {useV2: true})));
+        return of(0).pipe(delay(500), flatMap(() => this.apiService.post('wallet/seed/verify', {seed: seed}, {useV2: true})));
       } else {
         return of(0);
       }
