@@ -1,36 +1,42 @@
-import { browser, by, element } from 'protractor';
-
 export class SendPage {
   navigateTo() {
-    return browser.get('/#/send');
+    return browser.url('/#/send');
   }
 
   getHeaderText() {
-    return element(by.css('.title')).getText();
+    return $('.title').getText();
   }
 
-  getWalletsCount() {
-    return element.all(by.css('#wallet option')).count();
+  async getWalletsCount() {
+    await browser.keys("Escape");
+    
+    return $('.mat-mdc-select').click().then(() => {
+      return $$('.mat-mdc-select-panel mat-option span').length;
+    });
   }
 
-  getWalletsWithCoins() {
-    return element.all(by.tagName('#wallet option'))
-      .filter((opt) => {
+  async getWalletsWithCoins() {
+    await browser.keys("Escape");
+
+    return $('.mat-mdc-select').click().then(() => {
+      return $$('.mat-mdc-select-panel mat-option span').filter((opt) => {
         return opt.getText().then((v) => {
           return this.getCoinsFromOptionString(v) > 0;
         });
       });
+    });
   }
 
-  getValidWallets() {
-    return element.all(by.tagName('#wallet option'))
-      .filter((opt) => {
+  async getValidWallets() {
+    await browser.keys("Escape");
+
+    return $('.mat-mdc-select').click().then(() => {
+      return $$('.mat-mdc-select-panel mat-option span').filter((opt) => {
         return opt.getText().then((v) => {
-          return opt.getAttribute('disabled').then(status => {
-            return status === null && this.getCoinsFromOptionString(v) > 0;
-          });
+          return this.getCoinsFromOptionString(v) > 0;
         });
       });
+    });
   }
 
   selectValidWallet() {
@@ -41,18 +47,22 @@ export class SendPage {
     });
   }
 
-  fillFormWithCoins(coins: string) {
-    const dest = element(by.css('[formcontrolname="address"]'));
-    const amount = element(by.css('[formcontrolname="coins"]'));
-    const btnSend = element(by.buttonText('Send'));
+  async fillFormWithCoins(coins) {
+    const dest = $('[formcontrolname="address"]');
+    const amount = $('[formcontrolname="coins"]');
+    const btnSend = $("button=Send");
 
-    dest.clear();
-    amount.clear();
+    await dest.click;
+    await dest.clearValue();
+    await amount.click;
+    await amount.clearValue();
 
-    return dest.sendKeys('2e1erPpaxNVC37PkEv3n8PESNw2DNr5aJNy').then(() => {
+    await dest.click;
+    return dest.setValue('2e1erPpaxNVC37PkEv3n8PESNw2DNr5aJNy').then(() => {
       return this.getValidWallets().then(wallets => {
-        return wallets[0].click().then(() => {
-          return amount.sendKeys(coins).then(() => {
+        return wallets[0].click().then(async () => {
+          await amount.click;
+          return amount.setValue(coins).then(() => {
             return btnSend.isEnabled();
           });
         });
@@ -60,7 +70,7 @@ export class SendPage {
     });
   }
 
-  private getCoinsFromOptionString(option: string) {
+  getCoinsFromOptionString(option) {
     const value = option.slice(option.indexOf('-') + 1, option.indexOf(' SKY'));
 
     return parseFloat(value);
