@@ -17,6 +17,7 @@ import { OperationError, OperationErrorTypes } from '../../utils/operation-error
  */
 @Injectable()
 export class WalletsAndAddressesService {
+
   /**
    * Key used for saving the hw wallet list in persistent storage.
    */
@@ -450,7 +451,7 @@ export class WalletsAndAddressesService {
             // This is just a precaution.
             wallet.isHardware = true;
             if (!wallet.addresses) {
-              wallet.addresses = [{ address: 'invalid', confirmed: false, }];
+              wallet.addresses = [{ address: 'invalid', confirmed: false }];
             }
 
             wallet.id = wallet.addresses[0].address;
@@ -477,7 +478,7 @@ export class WalletsAndAddressesService {
    * @returns True if the address is valid or false otherwise.
    */
   verifyAddress(address: string): Observable<boolean> {
-    return this.apiService.post('address/verify', { address }, {useV2: true}).pipe(
+    return this.apiService.post('address/verify', { address: address }, {useV2: true}).pipe(
       map(() => true),
       catchError((err: OperationError) => {
         err = processServiceError(err);
